🚚 Move the SpannerStateTransaction for reuse
