+++ src/transaction/index.ts
@@ -1,2 +1,3 @@
 export * from './firestore-pubsub/index.js';
 export * from './spanner-pubsub/index.js';
+export { SpannerStateTransaction } from './spanner-state-transaction.js';
