+++ src/transaction/spanner-pubsub/index.ts
@@ -1,4 +1,3 @@
 export { SpannerPubSubTransactionModule } from './module.js';
 export { SpannerPubSubTransactionRunner } from './runner.js';
-export { SpannerStateTransaction } from './state-transaction.js';
 export { SpannerPubSubTransaction } from './transaction.js';
